BR-25 Fixed tests for "Package Scanner" * Tests structure has been changed
