--- src/test/java/com/bobocode/svydovets/beansForTestScan/SomeBean.java
@@ -1,7 +1,0 @@
-package com.bobocode.svydovets.beansForTestScan;
-
-import svydovets.core.annotation.Component;
-
-@Component
-public class SomeBean {
-}
