--- src/test/java/com/bobocode/svydovets/config/BeanConfigPostConstruct.java
@@ -1,9 +1,0 @@
-package com.bobocode.svydovets.config;
-
-import svydovets.core.annotation.ComponentScan;
-import svydovets.core.annotation.Configuration;
-
-@Configuration
-@ComponentScan("com.bobocode.svydovets.service.postconstruct")
-public class BeanConfigPostConstruct {
-}
