+++ src/test/java/com/bobocode/svydovets/ioc/core/beanDefinitionFactory/BeanDefinitionFactoryTest.java
@@ -1,0 +1,40 @@
+package com.bobocode.svydovets.ioc.core.beanDefinitionFactory;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+class BeanDefinitionFactoryTest {
+
+    @BeforeEach
+    void setUp() {
+    }
+
+    @AfterEach
+    void tearDown() {
+    }
+
+    @Test
+    void registerBeanDefinitions() {
+    }
+
+    @Test
+    void registerBeanDefinition() {
+    }
+
+    @Test
+    void getBeanDefinitionByBeanName() {
+    }
+
+    @Test
+    void createBeanDefinitionMapByConfigClass() {
+    }
+
+    @Test
+    void createComponentBeanDefinitionByBeanClass() {
+    }
+
+    @Test
+    void createBeanDefinitionByBeanInitMethod() {
+    }
+}
