+++ src/test/java/com/bobocode/svydovets/ioc/core/beanFactory/BeanFactoryTest.java
@@ -1,0 +1,36 @@
+package com.bobocode.svydovets.ioc.core.beanFactory;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+class BeanFactoryTest {
+
+    @BeforeEach
+    void setUp() {
+    }
+
+    @AfterEach
+    void tearDown() {
+    }
+
+    @Test
+    void registerBeans() {
+    }
+
+    @Test
+    void testRegisterBeans() {
+    }
+
+    @Test
+    void getBean() {
+    }
+
+    @Test
+    void testGetBean() {
+    }
+
+    @Test
+    void getBeansOfType() {
+    }
+}
