--- src/test/java/com/bobocode/svydovets/service/base/CommonService.java
@@ -1,7 +1,0 @@
-package com.bobocode.svydovets.service.base;
-
-import svydovets.core.annotation.Component;
-
-@Component
-public class CommonService {
-}
