--- src/test/java/com/bobocode/svydovets/service/base/NullService.java
@@ -1,7 +1,0 @@
-package com.bobocode.svydovets.service.base;
-
-import svydovets.core.annotation.Component;
-
-@Component
-public class NullService {
-}
