--- src/test/java/com/bobocode/svydovets/service/postconstruct/invalid/DuplicatePostConstructService.java
@@ -1,25 +1,0 @@
-package com.bobocode.svydovets.service.postconstruct.invalid;
-
-import svydovets.core.annotation.Component;
-import svydovets.core.annotation.PostConstruct;
-
-@Component
-public class DuplicatePostConstructService {
-
-    private String name;
-
-    @PostConstruct
-    private void setFirstName() {
-        this.name = "post-construct-service-1";
-    }
-
-    @PostConstruct
-    private void setSecondName() {
-        this.name = "post-construct-service-2";
-    }
-
-    public String getName() {
-        return name;
-    }
-
-}
