--- src/test/java/com/bobocode/svydovets/service/qualifier/NonPrimaryProductServiceImpl.java
@@ -1,7 +1,0 @@
-package com.bobocode.svydovets.service.qualifier;
-
-import svydovets.core.annotation.Component;
-
-@Component
-public class NonPrimaryProductServiceImpl implements ProductService {
-}
