--- src/test/java/com/bobocode/svydovets/service/qualifier/PrimaryProductServiceImpl.java
@@ -1,9 +1,0 @@
-package com.bobocode.svydovets.service.qualifier;
-
-import svydovets.core.annotation.Component;
-import svydovets.core.annotation.Primary;
-
-@Primary
-@Component
-public class PrimaryProductServiceImpl implements ProductService {
-}
