--- src/test/java/com/bobocode/svydovets/service/qualifier/ProductService.java
@@ -1,4 +1,0 @@
-package com.bobocode.svydovets.service.qualifier;
-
-public interface ProductService {
-}
