+++ src/test/java/com/bobocode/svydovets/source/autowire/method/TrimService.java
@@ -1,0 +1,20 @@
+package com.bobocode.svydovets.source.autowire.method;
+
+import com.bobocode.svydovets.source.autowire.field.EditService;
+import svydovets.core.annotation.Autowired;
+
+/**
+ * Class for testing "autowiring" logic via setter. Creates as bean in the {@link BeanConfig} class
+ */
+public class TrimService {
+    private EditService editService;
+
+    public String trimMessage() {
+        return editService.editMessage().trim();
+    }
+
+    @Autowired
+    public void setEditService(EditService editService) {
+        this.editService = editService;
+    }
+}
