+++ src/test/java/com/bobocode/svydovets/source/base/NullService.java
@@ -1,0 +1,7 @@
+package com.bobocode.svydovets.source.base;
+
+import svydovets.core.annotation.Component;
+
+@Component
+public class NullService {
+}
