+++ src/test/java/com/bobocode/svydovets/source/beansForTestScan/SomeBeanSecond.java
@@ -1,0 +1,7 @@
+package com.bobocode.svydovets.source.beansForTestScan;
+
+import svydovets.core.annotation.Component;
+
+@Component
+public class SomeBeanSecond {
+}
