+++ src/test/java/com/bobocode/svydovets/source/config/BeanConfigPostConstruct.java
@@ -1,0 +1,9 @@
+package com.bobocode.svydovets.source.config;
+
+import svydovets.core.annotation.ComponentScan;
+import svydovets.core.annotation.Configuration;
+
+@Configuration
+@ComponentScan("com.bobocode.svydovets.source.postconstruct")
+public class BeanConfigPostConstruct {
+}
