+++ src/test/java/com/bobocode/svydovets/source/config/MockConfig.java
@@ -1,0 +1,7 @@
+package com.bobocode.svydovets.source.config;
+
+import svydovets.core.annotation.Configuration;
+
+@Configuration
+public class MockConfig {
+}
