+++ src/test/java/com/bobocode/svydovets/source/qualifier/NonPrimaryProductServiceImpl.java
@@ -1,0 +1,7 @@
+package com.bobocode.svydovets.source.qualifier;
+
+import svydovets.core.annotation.Component;
+
+@Component
+public class NonPrimaryProductServiceImpl implements ProductService {
+}
