+++ src/test/java/com/bobocode/svydovets/source/qualifier/PrimaryProductServiceImpl.java
@@ -1,0 +1,9 @@
+package com.bobocode.svydovets.source.qualifier;
+
+import svydovets.core.annotation.Component;
+import svydovets.core.annotation.Primary;
+
+@Primary
+@Component
+public class PrimaryProductServiceImpl implements ProductService {
+}
