+++ src/test/java/com/bobocode/svydovets/source/qualifier/ProductService.java
@@ -1,0 +1,4 @@
+package com.bobocode.svydovets.source.qualifier;
+
+public interface ProductService {
+}
