+++ src/test/java/com/bobocode/svydovets/source/withoutAnnotation/NonAnnotatedClass.java
@@ -1,0 +1,4 @@
+package com.bobocode.svydovets.source.withoutAnnotation;
+
+public class NonAnnotatedClass {
+}
