+++ src/test/java/com/bobocode/svydovets/source/withoutAnnotation/NonAnnotatedClassFirst.java
@@ -1,0 +1,4 @@
+package com.bobocode.svydovets.source.withoutAnnotation;
+
+public class NonAnnotatedClassFirst {
+}
