--- src/test/java/com/bobocode/svydovets/withoutAnnotation/NonAnnotatedClassFirst.java
@@ -1,4 +1,0 @@
-package com.bobocode.svydovets.withoutAnnotation;
-
-public class NonAnnotatedClassFirst {
-}
