--- src/test/java/com/bobocode/svydovets/withoutAnnotation/NonAnnotatedClassSecond.java
@@ -1,4 +1,0 @@
-package com.bobocode.svydovets.withoutAnnotation;
-
-public class NonAnnotatedClassSecond {
-}
